% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchs.R
\name{matchs}
\alias{matchs}
\title{match region names}
\usage{
matchs(dataset, datacol, geofile, filecol)
}
\arguments{
\item{dataset}{data set with regions and values to be matched}

\item{datacol}{column values}

\item{geofile}{a geosjon file}

\item{filecol}{territory column to match}
}
\description{
mapping regions given geojson file
}
