% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmap.R
\name{regmap}
\alias{regmap}
\title{plot mapregions}
\usage{
regmap(geofile, territory, labels, measure, fill, label.size,main, mapcol, xlimit, ylimit)
}
\arguments{
\item{geofile}{geojon file with coordinates}

\item{territory}{regions}

\item{labels}{column values to be set in labels}

\item{measure}{column value to measure}

\item{fill}{type of label: 1-abrev, 2-full, 3- None}

\item{label.size}{size of labels}

\item{main}{main title}

\item{mapcol}{colores}

\item{xlimit}{x limit}

\item{ylimit}{y limit}
}
\description{
mapping regions given geojson file
}
