% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subst.R
\name{subst}
\alias{subst}
\title{substitute region names}
\usage{
subst(string, replace, dataset, datacol)
}
\arguments{
\item{string}{array with nonmatch values}

\item{replace}{regions replacement}

\item{dataset}{data set with regions and values to be matched}

\item{datacol}{column values}
}
\description{
mapping regions given geojson file
}
